Refactored the entire Undo system in several layers. + Added "undo" library (undo namespace). + Moved UndoHistory to undo namespace. + Added Undoer and UndoersCollector interfaces. + Converted old UndoStream to UndoersStack class. + Added new "undoers" namespace and moved each Undoer implementation   (classes to undo different actions) to "undoers" directory. + Added IO operations to "raster" objects (image, cel, layer, palette,   mask).

+++ src/raster/sprite.cpp
@@ -373,7 +373,7 @@ void Sprite::getCels(CelList& cels)
   getFolder()->getCels(cels);
 }
 
-void Sprite::remapImages(int frame_from, int frame_to, const std::vector<int>& mapping)
+void Sprite::remapImages(int frame_from, int frame_to, const std::vector<uint8_t>& mapping)
 {
   ASSERT(m_imgtype == IMAGE_INDEXED);
   ASSERT(mapping.size() == 256);
